Support for float story points when issue edit
On backlogs page predefined float value of story points is supported but on issue page not yet

+++ lib/backlogs_hooks.rb
@@ -122,7 +122,7 @@ module BacklogsPlugin
             if Backlogs.setting[:story_points].blank?
               snippet += context[:form].text_field(:story_points, :size => 3)
             else
-              snippet += context[:form].select(:story_points, options_for_select(Backlogs.setting[:story_points].split(',').map(&:to_i), issue.story_points.try(:to_i).try(:to_s)), include_blank: true)
+              snippet += context[:form].select(:story_points, options_for_select(Backlogs.setting[:story_points].split(',').map(&:to_f), issue.story_points.try(:to_f).try(:to_s)), include_blank: true)
             end
             snippet += '</p>'
